Reorganise Classes
