+++ examples/example.html
@@ -7,7 +7,7 @@
   <body>
     <canvas id="test"></canvas>
 
-    <script src="../dist/index.js"></script>
+    <script src="../dist/CanvasManager.js"></script>
 
     <script type="application/javascript">
       const manager = new window.CanvasManager(document.querySelector('#test'));
